feat(ui): split main script into modules
